Move all new code for evaluation to Engine module
