+++ rvrs-lang.cabal
@@ -15,6 +15,7 @@ library
     , RVRS.Parser.ExprParser
     , RVRS.Parser.StmtParser
     , RVRS.Parser.Type
+    , RVRS.Engine
     , RVRS.Eval
     , RVRS.Pretty
     , RVRS.Value
@@ -55,6 +56,7 @@ executable rvrs
     , RVRS.Parser.ExprParser
     , RVRS.Parser.StmtParser
     , RVRS.Parser.Type
+    , RVRS.Engine
     , RVRS.Eval
     , RVRS.Pretty
     , RVRS.Value
@@ -98,6 +100,7 @@ executable RunAll
     , RVRS.Parser.ExprParser
     , RVRS.Parser.StmtParser
     , RVRS.Parser.Type
+    , RVRS.Engine
     , RVRS.Eval
     , RVRS.Pretty
     , RVRS.Value
@@ -144,6 +147,7 @@ executable TestLower
     , RVRS.Parser.ExprParser
     , RVRS.Parser.StmtParser
     , RVRS.Parser.Type
+    , RVRS.Engine
     , RVRS.Eval
     , RVRS.Pretty
     , RVRS.Value
@@ -187,6 +191,7 @@ executable RunIRTests
     , RVRS.Parser.ExprParser
     , RVRS.Parser.StmtParser
     , RVRS.Parser.Type
+    , RVRS.Engine
     , RVRS.Eval
     , RVRS.Pretty
     , RVRS.Value
